requirements.txt
cryptography==3.4.7
Pillow==9.0.1
numpy==1.19.5